0.002
Readme changes

corrections

comment typo

"code" -> "expr"

.h -> hpp

echo

changed formatter to return std::string

updated formatter

unified default formatter, channels, & emit thresholds

moved default formatter into config.

fixed enum derps, renamed channel_enabled

impl echo test

ECHO and ECHOV

+ eofbit exception

removed unused headers

updated documentation

echo

comment

typo; how the hell was this compiling?

disabled enumerate test (libc bug)
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100639

impl czip tests

+++ include/fgl/utility/zip.hpp
@@ -164,7 +164,9 @@ template<std::ranges::forward_range ... T_ranges>
 constexpr auto zip(T_ranges&& ... args)
 {
 	using fgl::internal::forward_zip_iterator;
-	const zip_sentinel_t shortest{ std::min({ std::ranges::ssize(args)... }) };
+	const zip_sentinel_t shortest{
+		std::min({static_cast<std::ptrdiff_t>(std::ranges::ssize(args))...})
+	};
 	return fgl::range_wrapper(
 		forward_zip_iterator(std::begin(std::forward<T_ranges>(args))...),
 		shortest
@@ -177,7 +179,9 @@ template<std::ranges::forward_range ... T_ranges>
 constexpr auto czip(T_ranges&& ... args)
 {
 	using fgl::internal::forward_zip_iterator;
-	const zip_sentinel_t shortest{ std::min({ std::ranges::ssize(args)... }) };
+	const zip_sentinel_t shortest{
+		std::min({static_cast<std::ptrdiff_t>(std::ranges::ssize(args))...})
+	};
 	return fgl::range_wrapper(
 		forward_zip_iterator(std::cbegin(std::forward<T_ranges>(args))...),
 		shortest
